% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataR.R
\docType{data}
\name{dataR}
\alias{dataR}
\title{Presence-absence of 33 Australian tree frogs (Rosauer 2017)}
\format{
A matrix with 2891 rows and 35 columns.
}
\source{
Rosauer, 2017. Available on:
\url{https://github.com/DanRosauer/phylospatial/tree/master/PhyloEndemism_in_R/Tree\%20Frog\%20Data}
}
\usage{
dataR
}
\description{
A dataset containing presence-absence of 33 Australian tree frogs.
This dataset also provide coordinates x and y for each site.
}
\keyword{datasets}
