% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2rast.R
\name{df2rast}
\alias{df2rast}
\title{Transform data.frame to raster}
\usage{
df2rast(x)
}
\arguments{
\item{x}{data.frame. A data.frame or matrix with species names in columns and sites in rows. The first two columns must provide x and y coordinates.}
}
\value{
SpatRaster
}
\description{
Transform data.frame to raster
}
\examples{
\dontrun{
dat <- phylogrid::dataR
df.to.rast(dat)
}
}
