% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{.fit.memory}
\alias{.fit.memory}
\title{Evaluate if the rasters generated in the function fits on available memory}
\usage{
.fit.memory(x, n = 1)
}
\arguments{
\item{x}{SpatRaster}

\item{n}{positive integer. The number of copies of \code{x} that are needed}
}
\value{
logical
}
\description{
Tests if the amount of RAM required is available to process a SpatRaster
}
\keyword{internal}
