% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.ed.ses.R
\name{.rast.ed.B}
\alias{.rast.ed.B}
\title{Calculate Evolutionary distinctiveness for each raster cell}
\usage{
.rast.ed.B(x, branch.length, n.descen, cores = 1, filename = NULL, ...)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species. The layers (species) must be sorted according to the tree order. See the phylo.pres function.}

\item{branch.length}{numeric. A Named numeric vector containing the branch length of each specie.}

\item{n.descen}{numeric. A Named numeric vector of number of descendants for each branch}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{filename}{character. Output filename.}

\item{...}{additional arguments to be passed passed for fun.}
}
\value{
SpatRaster
}
\description{
This function calculates evolutionary distinctiveness according to the fair-proportion index.
}
\references{
Isaac, N. J., Turvey, S. T., Collen, B., Waterman, C. and Baillie, J. E. (2007). Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE 2, e296.
}
\author{
Gabriela Alves-Ferreira and Neander Marcel Heming
}
