% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pe.ses.R
\name{.rast.pe.B}
\alias{.rast.pe.B}
\title{Calculate phylogenetic endemism (PE. Rosauer et al. 2009) for a raster}
\usage{
.rast.pe.B(x, branch.length, filename = NULL, ...)
}
\arguments{
\item{x}{SpatRaster. A presence-absence SpatRaster with the layers ordered according to the tree order}

\item{branch.length}{numeric. A Named numeric vector containing the branch length of each specie}

\item{filename}{character. Output filename.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculate phylogenetic endemism following Rosauer et al. (2009) using rasters as input and output.
}
\references{
Rosauer, D. A. N., Laffan, S. W., Crisp, M. D., Donnellan, S. C., & Cook, L. G. (2009). Phylogenetic endemism: a new approach for identifying geographical concentrations of evolutionary history. Molecular ecology, 18(19), 4061-4072.
}
