% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{.vec.geo.phylo2}
\alias{.vec.geo.phylo2}
\title{Calculate phylogenetic community metrics for each raster cell}
\usage{
.vec.geo.phylo2(
  xinv.R,
  branch.length,
  resu = stats::setNames(as.double(rep(NA, 2)), c("PE", "WE"))
)
}
\arguments{
\item{xinv.R}{SpatRaster. multiplication of x and inv.R from \code{\link{.rast.geo.phylo}}}

\item{branch.length}{numeric. A named numerical vector containing the branch
length for each species.}

\item{resu}{vector structuring SpatRaster result}
}
\value{
numeric
}
\description{
Calculate phylogenetic endemism and weighted endemism using vectors as input.
}
\author{
Neander Marcel Heming
}
