% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pe.ses.R
\name{.vec.pe}
\alias{.vec.pe}
\title{Calculate phylogenetic endemism for a vector}
\usage{
.vec.pe(x, branch.length, resu = as.double(NA))
}
\arguments{
\item{x}{numeric. A named numerical vector of presence-absence for one sample.}

\item{branch.length}{numeric. A named numerical vector containing the branch
length for each species.}

\item{resu}{vector structuring SpatRaster result}
}
\value{
numeric
}
\description{
This function calculates phylogenetic endemism for a vector
}
\references{
Rosauer, D. A. N., Laffan, S. W., Crisp, M. D., Donnellan, S. C.,
& Cook, L. G. (2009). Phylogenetic endemism: a new approach for identifying
geographical concentrations of evolutionary history. Molecular ecology,
18(19), 4061-4072.
}
\author{
Neander Marcel Heming and Gabriela Alves-Ferreira
}
