% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pe.ses.R
\name{.vec.wpe}
\alias{.vec.wpe}
\title{Calculate phylogenetic endemism for a vector}
\usage{
.vec.wpe(x, spp_seq, spp_seq2, wpe = as.double(NA))
}
\arguments{
\item{x}{numeric. A named numerical vector of presence-absence for one sample.}

\item{spp_seq}{numeric vector indicating presence/absence data in 'x'}

\item{spp_seq2}{numeric. numeric vector indicating presence/absence data in 'x'}

\item{wpe}{numeric. numeric vector indicating INV data (inverse range size) in 'x'}
}
\value{
numeric
}
\description{
This function calculates phylogenetic endemism for a vector
}
\references{
Rosauer, D. A. N., Laffan, S. W., Crisp, M. D., Donnellan, S. C.,
& Cook, L. G. (2009). Phylogenetic endemism: a new approach for identifying
geographical concentrations of evolutionary history. Molecular ecology,
18(19), 4061-4072.
}
\author{
Neander Marcel Heming and Gabriela Alves-Ferreira
}
