% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.rosauer.R
\name{load.data.rosauer}
\alias{load.data.rosauer}
\title{Load an example dataset with presence-absence data of 33 tree frogs and a phylogenetic tree for this species}
\source{
Rosauer, 2017. Available on: \url{https://github.com/DanRosauer/phylospatial/tree/master/PhyloEndemism_in_R/Tree\%20Frog\%20Data}
}
\usage{
load.data.rosauer()
}
\value{
data.frame and phylo
}
\description{
This function load a phylogenetic tree and a dataset with presence-absence of
33 Australian tree frogs from Rosauer (2017). Also provides coordinates to each site.
}
