% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.pres.R
\name{phylo.pres}
\alias{phylo.pres}
\title{Reorder stack according to tree order and get branch length}
\usage{
phylo.pres(pres.stack, tree)
}
\arguments{
\item{pres.stack}{a raster of presence-absence of class 'SpatRaster'}

\item{tree}{an object of class 'phylo'}
}
\value{
SpatRaster and numeric vector
}
\description{
Reorder stack according to tree order and get branch length.
}
