% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\title{Calculate range size}
\usage{
range(x, scale = FALSE)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species. This raster must be named.}

\item{scale}{logical. If TRUE, scaling is done by dividing the range size x by the area total. If FALSE, scaling is not done. The default is FALSE.}
}
\value{
vector
}
\description{
This function calculate range size in square kilometers for all cells that are not NA. The size of the cells is constant in degrees but not in square meters, which was considered in the method applied to calculate the area. If scale is TRUE, the raster is scaled from 0 to 1.
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
range(ras, scale = TRUE)
}
}
\author{
Gabriela Alves Ferreira and Neander Marcel Heming
}
