% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_size.R
\name{range_size}
\alias{range_size}
\title{Calculate range size for a set of species using a raster as imput}
\usage{
range_size(x, unit = "m", scale = FALSE, ...)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species. This raster must be named.}

\item{unit}{character. One of "m", "km", or "ha"}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
vector
}
\description{
This function calculate range size in square meters (by default) for all cells that are not NA. The size of the cells is constant in degrees but not in square meters, which was considered in the method applied to calculate the area.
}
\examples{
\dontrun{
x <- terra::rast(system.file("extdata", "rast.presab.tif", package="phyloraster"))
range_size(x)
}
}
\author{
Gabriela Alves Ferreira and Neander Marcel Heming
}
