% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{rast.se}
\alias{rast.se}
\title{Calculate species richness for each raster cell}
\usage{
rast.se(x, filename = NULL, cores = 1, ...)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species.}

\item{filename}{character. Output filename.}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculate the species richness for each raster cell.
}
\examples{
\dontrun{
x <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
rse <- rast.se(x)
terra::plot(rse)
}
}
\author{
Gabriela Alves Ferreira and Neander Marcel Heming
}
