% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.we.ses.R
\name{rast.we.ses}
\alias{rast.we.ses}
\title{Calculate Weighted Endemism (WE. Williams et al. 1994, Crisp et al. 2001) standardized for species richness}
\usage{
rast.we.ses(
  x,
  aleats,
  random = c("area.size", "site", "specie", "fullspat"),
  filename = NULL
)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster of presence-absence.}

\item{aleats}{numeric. A number indicating how many times the calculation should be repeated.}

\item{random}{character. A character indicating what type of randomization. Could be by tip, site, specie or full spat(site and specie).}

\item{filename}{character. Output filename.}
}
\value{
SpatRaster
}
\description{
The function calculates the WE corrected for species richness. In each null model run, richness is kept constant and range size of each species are randomized. The function provides the mean, standard deviation of all null models and also calculates the standardized effect size (SES).
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
t <- rast.we.ses(ras, aleats = 10)
}
}
\references{
Williams, P.H., Humphries, C.J., Forey, P.L., Humphries, C.J., VaneWright, R.I. (1994). Biodiversity, taxonomic relatedness, and endemism in conservation. In: Systematics and Conservation Evaluation (eds Forey PL, Humphries CJ, Vane-Wright RI), p. 438. Oxford University Press, Oxford.

Crisp, M., Laffan, S., Linder, H., Monro, A. (2001). Endemism in theAustralian flora. Journal of Biogeography, 28, 183–198.
}
