% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_pd.R
\name{rast_pd}
\alias{rast_pd}
\title{Calculate phylogenetic diversity for a raster of presence-absence}
\usage{
rast_pd(pres_bin_stack, tree)
}
\arguments{
\item{pres_bin_stack}{a raster of presence-absence. It can be an object of class 'raster' or 'SpatRaster'}

\item{tree}{an object of class 'phylo'}
}
\value{
SpatRaster
}
\description{
Calculate phylogenetic diversity for a raster of presence-absence
}
