% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.rand.R
\name{spat.rand}
\alias{spat.rand}
\title{Randomize a set of rasters according to the observed frequency.}
\usage{
spat.rand(x, aleats, random = c("site", "species", "fullspat"), cores = 1)
}
\arguments{
\item{x}{SpatRaster. A presence-absence SpatRaster.}

\item{aleats}{numeric. A number indicating how many times the calculation should be repeated.}

\item{random}{character. Character indicating the type of randomization to be used. The available types are by "site", "specie" or "fullspat".}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}
}
\value{
SpatRaster
}
\description{
Randomize a set of rasters according to the observed frequency using the methods: sites (by cells), species (by layer) or both (layers and cells).
}
\examples{
\dontrun{
aleats = 10
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
sr <- spat.rand(ras, aleats = 10, random = "site")
plot(sr)
}
}
\author{
Neander Marcel Heming and Gabriela Alves-Ferreira
}
