% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.length.R
\name{species.branch.length}
\alias{species.branch.length}
\title{Compute species branch lengths}
\usage{
species.branch.length(tree, edge.info = NULL, ...)
}
\arguments{
\item{tree}{phylo. A dated tree.}

\item{edge.info}{Object returned by \code{\link{tip.root.path}} consisting of
a list containing the edge matrix (H1) with the path from tip to root and and
a numeric vector (edge.length) giving the length of each branch of the tree.}

\item{...}{additional arguments to be passed passed down from a calling
function.}
}
\value{
returns a numeric vector giving the length of species branch.
}
\description{
Computation of species branch lengths using a phylogeny
}
\details{
Calculates species branch lengths using the edge length and edge path
across the phylogeny
}
\examples{
library(phyloraster)
tree <- ape::read.tree(system.file("extdata", "tree.nex",
package="phyloraster"))

species.branch.length(tree)

library(ape)
set.seed(1)

tree <- rtree(n=40)

plot(tree)

species.branch.length(tree)

#' library(ape)
set.seed(1)

tree <- rtree(n=40)

plot(tree)

edge.info <- tip.root.path(tree)

species.branch.length(edge.info = edge.info)

}
\author{
Neander M. Heming
}
