% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.rand.R
\name{tip.rand}
\alias{tip.rand}
\title{Randomize branch lengths for x times}
\usage{
tip.rand(x, aleats)
}
\arguments{
\item{x}{numeric. A numeric vector.}

\item{aleats}{numeric. A number indicating how many times the calculation should be repeated.}
}
\value{
list
}
\description{
This function randomize branch lengths or another vector for a determined number of aleatorizations.
}
\author{
Gabriela Alves-Ferreira
}
