% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2rast.R
\name{df2rast}
\alias{df2rast}
\title{Transform a data.frame to raster}
\usage{
df2rast(x, CRS)
}
\arguments{
\item{x}{data.frame. A data.frame or matrix with species names in columns and sites in rows. The first two columns must provide longitude and latitude, respectively.}

\item{CRS}{character. A coordinate reference system (projection) for a Raster object.}
}
\value{
SpatRaster
}
\description{
The function transform a data.frame or a matrix of presence-absence in a raster of distribution.
}
\examples{
\dontrun{
dat <- phylogrid::load.data.rosauer()
df2rast(dat$presab, CRS = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
}
}
