% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2rast.R
\name{df2rast}
\alias{df2rast}
\title{Transform a data.frame to raster}
\usage{
df2rast(x, ...)
}
\arguments{
\item{x}{data.frame. A data.frame or matrix with species names in columns and sites in rows. The first two columns must provide longitude and latitude, respectively.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
The function transform a data.frame or a matrix of presence-absence in a raster of distribution.
}
\examples{
\dontrun{
dat <- phylogrid::load.data.rosauer()
df2rast(dat$presab)
}
}
