% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2rast.R
\name{df2rast}
\alias{df2rast}
\title{Transform data.frame to raster}
\usage{
df2rast(x, resolution = NULL)
}
\arguments{
\item{x}{data.frame. A data.frame or matrix with species names in columns and sites in rows. The first two columns must provide x and y coordinates.}

\item{CRS}{character. A coordinate reference system (projection) for a Raster object.}
}
\value{
SpatRaster
}
\description{
Transform data.frame to raster
}
\examples{
\dontrun{
dat <- phylogrid::load.data.Rosauer
df2rast(dat, CRS = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
}


}
