% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pe.ses.R
\name{.rast.pe.B}
\alias{.rast.pe.B}
\title{Calculate phylogenetic endemism for a raster}
\usage{
.rast.pe.B(x, spp_seq, spp_seqrange.BL, cores = 1, filename = "", ...)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1)
for a set of species. The layers (species) must be sorted according to the
tree order. See the phylo.pres function.}

\item{spp_seq}{numeric vector indicating presence/absence data in 'x'}

\item{spp_seqrange.BL}{numeric vector indicating range.BL data
(inverse range size * branch lengh) in 'x'}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with
that many cores is created and used.}

\item{filename}{character. Output filename.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculate phylogenetic endemism using rasters as input and output.
}
\references{
Rosauer, D. A. N., Laffan, S. W., Crisp, M. D., Donnellan, S. C.,
& Cook, L. G. (2009). Phylogenetic endemism: a new approach for identifying
geographical concentrations of evolutionary history. Molecular ecology,
18(19), 4061-4072.
}
