% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.we.ses.R
\name{.rast.we.B}
\alias{.rast.we.B}
\title{Calculate weighted endemism for each raster cell}
\usage{
.rast.we.B(x, filename = NULL, cores = 1, ...)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species.}

\item{filename}{character. Output filename.}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculate the sum of the inverse of the range size for species present in each raster cell.
}
\examples{
\dontrun{
x <- terra::rast(system.file("extdata", "rast.presab.tif", package="phyloraster"))
rs <- range_size(x)
.rast.we.B(x, rs)
}

}
\references{
Williams, P.H., Humphries, C.J., Forey, P.L., Humphries, C.J., VaneWright, R.I. (1994). Biodiversity, taxonomic relatedness, and endemism in conservation. In: Systematics and Conservation Evaluation (eds Forey PL, Humphries CJ, Vane-Wright RI), p. 438. Oxford University Press, Oxford.

Crisp, M., Laffan, S., Linder, H., Monro, A. (2001). Endemism in theAustralian flora. Journal of Biogeography, 28, 183–198.
}
\author{
Neander Marcel Heming and Gabriela Alves-Ferreira
}
