% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.ed.ses.R
\name{.vec.ed}
\alias{.vec.ed}
\title{Calculate Evolutionary Distinctiveness for a vector}
\usage{
.evol.distin(x, branch.length, n.descen)
}
\arguments{
\item{x}{numeric. A Named numeric vector of presence-absence}

\item{branch.length}{numeric. A Named numeric vector of branch length for
each species}

\item{n.descen}{numeric. A Named numeric vector of number of descendants for
each branch}
}
\value{
numeric
}
\description{
This function calculates evolutionary distinctiveness for a set
of species using the fair-proportion index (Isaac et al., 2007).
}
\references{
Isaac, N. J., Turvey, S. T., Collen, B., Waterman, C. and
Baillie, J. E. (2007). Mammals on the EDGE: conservation priorities based on
threat and phylogeny. PLoS ONE 2, e296.
}
\author{
Gabriela Alves-Ferreira and Neander Marcel Heming
}
