% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{.vec.geo.phylo}
\alias{.vec.geo.phylo}
\title{Calculate phylogenetic community metrics for each raster cell}
\usage{
.vec.geo.phylo(
  x,
  branch.length,
  n.descen,
  spp_seq,
  spp_seqLR,
  spp_seqINV,
  resu
)
}
\arguments{
\item{x}{numeric vector of presence/absence (1/0) data}

\item{branch.length}{numeric. A named numerical vector containing the branch
length for each species.}

\item{n.descen}{numeric. A Named numeric vector of number of descendants for
each branch}

\item{spp_seq}{numeric vector indicating presence/absence data in 'x'}

\item{spp_seqLR}{numeric vector indicating LR data (inverse range size * branch lengh) in 'x'}

\item{spp_seqINV}{numeric vector indicating INV data (inverse range size) in 'x}
}
\value{
numeric
}
\description{
Calculate species richness, phylogenetic diversity, evolutionary distinctiveness,
phylogenetic endemism and weighted endemism using vectors as input.
}
\author{
Neander Marcel Heming
}
