% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{.vec.geo.phylo}
\alias{.vec.geo.phylo}
\title{Calculate phylogenetic community metrics for each raster cell}
\usage{
.vec.geo.phylo(
  x,
  branch.length,
  n.descen,
  resu = stats::setNames(as.double(rep(NA, 3)), c("SR", "PD", "ED"))
)
}
\arguments{
\item{x}{numeric. A named numerical vector of presence-absence for one sample.}

\item{branch.length}{numeric. A named numerical vector containing the branch
length for each species.}

\item{n.descen}{numeric. A Named numeric vector of number of descendants for
each branch}

\item{resu}{vector structuring SpatRaster result}
}
\value{
numeric
}
\description{
Calculate species richness, phylogenetic diversity, and evolutionary
distinctiveness using vectors as input.
}
\author{
Neander Marcel Heming
}
