% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{.vec.pd}
\alias{.vec.pd}
\title{Calculate phylogenetic diversity (Faith 1992) for a vector}
\usage{
.vec.pd(pres.ab, branch.length)
}
\arguments{
\item{pres.ab}{numeric. A Named num vector of presence-absence}

\item{branch.length}{numeric. A Named num vector of branch length for each specie}
}
\value{
numeric
}
\description{
Calculate phylogenetic diversity (Faith 1992) for a vector
}
