% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pd.ses.R
\name{.vec.pd}
\alias{.vec.pd}
\title{Calculate phylogenetic diversity (Faith 1992) for a vector}
\usage{
.vec.pd(x, branch.length, pd = c(PD = NA))
}
\arguments{
\item{x}{numeric. A named numerical vector of presence-absence for one sample.}

\item{branch.length}{numeric. A named numerical vector containing the branch
length for each species.}

\item{pd}{numeric. A numeric vector with the values of phylogenetic diversity.}
}
\value{
numeric
}
\description{
This function calculates the sum of the branch length for a set
of species for one sample.
}
\references{
Faith, D. P. (1992). Conservation evaluation and phylogenetic
diversity. Biological conservation, 61(1), 1-10.
}
\author{
Neander Marcel Heming and Gabriela Alves-Ferreira
}
