% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pe.ses.R
\name{.vec.wpe}
\alias{.vec.wpe}
\title{Calculate phylogenetic endemism for a vector}
\usage{
.vec.wpe(x, spp_seq, spp_seq2, wpe = as.double(NA))
}
\arguments{
\item{x}{numeric. A named numerical vector of presence-absence for one sample.}

\item{branch.length}{numeric. A named numerical vector containing the branch length for each species.}
}
\value{
numeric
}
\description{
This function calculates phylogenetic endemism for a vector
}
\references{
Faith, D. P. (1992). Conservation evaluation and phylogenetic diversity. Biological conservation, 61(1), 1-10.
}
\author{
Neander Marcel Heming and Gabriela Alves-Ferreira
}
