% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{geo.phylo}
\alias{geo.phylo}
\title{Calculate phylogenetic diversity, phylogenetic endemism and weighted endemism for a raster}
\usage{
geo.phylo(x, branch.length, filename = NULL, cores = 1, ...)
}
\arguments{
\item{x}{SpatRaster. A presence-absence SpatRaster with the layers ordered according to the tree order.}

\item{branch.length}{numeric. A Named numeric vector containing the branch length of each specie.}

\item{filename}{character. Output filename.}

\item{cores}{numeric. A positive integer indicating the number of clusters used for parallelization}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculate phylogenetic diversity, phylogenetic endemism and weighted endemism using rasters as input and output.
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phylogrid"))
data <- phylo.pres(ras, tree)
geo.phylo(data$x, data$branch.length, cores = 1)
}
}
\references{
Rosauer, D. A. N., Laffan, S. W., Crisp, M. D., Donnellan, S. C., & Cook, L. G. (2009). Phylogenetic endemism: a new approach for identifying geographical concentrations of evolutionary history. Molecular ecology, 18(19), 4061-4072.

Faith, D. P. (1992). Conservation evaluation and phylogenetic diversity. Biological conservation, 61(1), 1-10.

Williams, P.H., Humphries, C.J., Forey, P.L., Humphries, C.J., VaneWright, R.I. (1994). Biodiversity, taxonomic relatedness, and endemism in conservation. In: Systematics and Conservation Evaluation (eds Forey PL, Humphries CJ, Vane-Wright RI), p. 438. Oxford University Press, Oxford.

Crisp, M., Laffan, S., Linder, H., Monro, A. (2001). Endemism in theAustralian flora. Journal of Biogeography, 28, 183–198.
}
