% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_pd.R
\name{geo.phylo}
\alias{geo.phylo}
\title{Calculate PD, PE and WE for a raster
Calculate phylogenetic diversity, phylogenetic endemism and weighted endemism using rasters as input and output}
\usage{
geo.phylo(pres.reord, area.inv, area.tips, branch.length, filename = NULL)
}
\arguments{
\item{pres.reord}{a presence-absence SpatRaster with the layers ordered according to the tree order}

\item{area.inv}{a presence-absence SpatRaster with the inverse of the area for each specie}

\item{area.tips}{a presence-absence SpatRaster with the inverse of the area vs branch length}

\item{branch.length}{a numeric vector containing the branch length of each specie}

\item{filename}{a character specifying the path where rasters can be saved}
}
\value{
SpatRaster
}
\description{
Calculate PD, PE and WE for a raster
Calculate phylogenetic diversity, phylogenetic endemism and weighted endemism using rasters as input and output
}
