% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{geo.phylo2}
\alias{geo.phylo2}
\title{Calculate PD, PE and WE for a raster- version 2}
\usage{
geo.phylo2(rast.branch, range.branch, filename = NULL)
}
\arguments{
\item{rast.branch}{SpatRaster. a presence-absence SpatRaster with the layers ordered according to the tree order and a vector with the branch lengths. These objects can be obtained in the phylo.pres function}

\item{range.branch}{SpatRaster. a presence-absence SpatRaster with the inverse of the area and the inverse of the area vs branch length. These rasters can be obtained in the inv.range function}

\item{filename}{character. Output filename}
}
\value{
SpatRaster
}
\description{
Calculate phylogenetic diversity, phylogenetic endemism and weighted endemism using rasters as input and output
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phylogrid"))
data <- phylo.pres(ras, tree)
range <- inv.range(data$pres.reord, data$branch.length)
geo.phylo2(data, range)
}
}
\references{
Rosauer, D. A. N., Laffan, S. W., Crisp, M. D., Donnellan, S. C., & Cook, L. G. (2009). Phylogenetic endemism: a new approach for identifying geographical concentrations of evolutionary history. Molecular ecology, 18(19), 4061-4072.

Faith, D. P. (1992). Conservation evaluation and phylogenetic diversity. Biological conservation, 61(1), 1-10.

Williams, P.H., Humphries, C.J., Forey, P.L., Humphries, C.J., VaneWright, R.I. (1994). Biodiversity, taxonomic relatedness, and endemism in conservation. In: Systematics and Conservation Evaluation (eds Forey PL, Humphries CJ, Vane-Wright RI), p. 438. Oxford University Press, Oxford.

Crisp, M., Laffan, S., Linder, H., Monro, A. (2001). Endemism in theAustralian flora. Journal of Biogeography, 28, 183–198.
}
