% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv.range.R
\name{inv.range}
\alias{inv.range}
\title{Calculate the inverse of range size}
\usage{
inv.range(x, branch.length = NULL, cores = 1, ...)
}
\arguments{
\item{x}{SpatRaster. A presence-absence SpatRaster with the layers ordered according to the tree order.}

\item{branch.length}{numeric. A vector containing the branch length of each specie.}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{...}{additional arguments to be passed passed down from a calling function.}

\item{LR}{logical. If LR = TRUE, the function will returns a raster with the clade range.}

\item{filename}{character. Output filename.}
}
\value{
SpatRaster and numeric
}
\description{
Get range size in square kilometers for all cells that are not NA, the inverse of range size and the inverse of range size multiplied by branch length for multiple species using a raster of presence-absence.
}
\examples{
x <- terra::rast(system.file("extdata", "rast.presab.tif", package="phyloraster"))
inv.range(x)

}
\author{
Neander Marcel Heming and Gabriela Alves-Ferreira
}
