% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.pres.R
\name{phylo.pres}
\alias{phylo.pres}
\title{Reorder stack according to tree order and get branch length}
\usage{
phylo.pres(x, tree)
}
\arguments{
\item{x}{SpatRaster. A raster of presence-absence}

\item{tree}{phylo. A dated tree}
}
\value{
SpatRaster, phylo and numeric vector
}
\description{
Reorder stack according to tree order and get branch length
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phylogrid"))
phylo.pres(ras, tree)
}
}
