% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.pres.R
\name{phylo.pres}
\alias{phylo.pres}
\title{Prepare rasters and phylogenetic tree to run community metrics}
\usage{
phylo.pres(x, tree, ...)
}
\arguments{
\item{x}{SpatRaster. A stack containing binary presence-absence rasters for each species.}

\item{tree}{phylo. A dated tree.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
Returns a list containing a SpatRaster reordered according to the order that the species appear in the phylogenetic tree, a subtree containing only the species that are in the stack of rasters and finally two named numerical vectors containing the branch length and the number of descendants of each species.
}
\description{
Reorder stack according to tree order and get branch length for each species. The names must be the same in the phylogenetic tree and in the raster for the same species. For example, if you have the name "Leptodactylus_latrans" in the raster and "Leptodactylus latrans" in the tree, the function will not work. The same goes for uppercase and lowercase letters.
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phylogrid"))
phylo.pres(ras, tree)
}
}
\author{
Neander Marcel Heming and Gabriela Alves Ferreira
}
