% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_pd.R
\name{phylo.pres}
\alias{phylo.pres}
\title{Data preparation}
\usage{
phylo.pres(pres.stack, tree)
}
\arguments{
\item{pres.stack}{a raster of presence-absence. It can be an object of class 'raster' or 'SpatRaster'}

\item{tree}{an object of class 'phylo'}
}
\value{
SpatRaster and numeric vector
}
\description{
Reorder stack according to tree order, get branch length, inverse area, and inverse area x branch lengths.
}
