% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{rast.ed}
\alias{rast.ed}
\title{Calculate Evolutionary distinctiveness for each raster cell}
\usage{
rast.ed(x, branch.length, n.descen, cores = 1, filename = NULL, ...)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species. The layers (species) must be sorted according to the tree order. See the phylo.pres function.}

\item{branch.length}{numeric. A Named numeric vector containing the branch length of each specie.}

\item{n.descen}{numeric. A Named numeric vector of number of descendants for each branch}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{filename}{character. Output filename.}

\item{...}{additional arguments to be passed passed for fun.}
}
\value{
SpatRaster
}
\description{
This function calculates evolutionary distinctiveness according to the fair-proportion index.
}
\examples{
\dontrun{
x <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
# phylogenetic tree
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phylogrid"))
data <- phylogrid::phylo.pres(x, tree)
ed <- phylogrid::rast.ed(data$x, data$branch.length, data$n.descen, cores = 2)
}
}
\references{
Isaac, N. J., Turvey, S. T., Collen, B., Waterman, C. and Baillie, J. E. (2007). Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE 2, e296.
}
\author{
Gabriela Alves-Ferreira and Neander Marcel Heming
}
