% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.ed.ses.R
\name{rast.ed.ses}
\alias{rast.ed.ses}
\title{Evolutionary distinctiveness standardized for specie richness}
\usage{
rast.ed.ses(
  x,
  branch.length,
  n.descen,
  aleats,
  random = c("tip", "spat"),
  cores = 1,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species.}

\item{branch.length}{numeric. A named numerical vector containing the branch length of each species. See phylo.pres function.}

\item{n.descen}{numeric. A Named numeric vector of number of descendants for each branch}

\item{aleats}{positive integer. A positive integer indicating how many times the calculation should be repeated.}

\item{random}{character. A character indicating the type of randomization. The currently available randomization methods are "tip", "site", "species" or "both" (site and species).}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{filename}{character. Output filename.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculates the standardized effect size for evolutionary distinctiveness. The function has four different methods for spatial and phylogenetic randomization. See Details for more information.
}
\details{
The "tip" method shuffles the taxon names among all those in the phylogeny. The "site" method keeps species richness constant in each pixel, but randomizes the position of species in the stack. In the "species" method, the order of species in the stack is kept constant, but the pixels where each species is present are randomized in space. The third method, "full.spat", combines site and species randomization at the same time.
}
\examples{
\dontrun{
x <- terra::rast(system.file("extdata", "rast.presab.tif", package="phyloraster"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phyloraster"))
data <- phyloraster::phylo.pres(x, tree)
t <- rast.ed.ses(data$x, data$branch.length, data$n.descendants, aleats = 5,
                 random = "both")
terra::plot(t)
}
}
\references{
Isaac, N. J., Turvey, S. T., Collen, B., Waterman, C. and Baillie, J. E. (2007). Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE 2, e296.

Laffan, S. W., Rosauer, D. F., Di Virgilio, G., Miller, J. T., González‐Orozco, C. E., Knerr, N., ... & Mishler, B. D. (2016). Range‐weighted metrics of species and phylogenetic turnover can better resolve biogeographic transition zones. Methods in Ecology and Evolution, 7(5), 580-588.
}
\author{
Gabriela Alves-Ferreira and Neander Heming
}
