% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{rast.pd}
\alias{rast.pd}
\title{Calculate phylogenetic diversity (PD. Faith, 1992) for a raster}
\usage{
rast.pd(x, branch.length, filename = NULL, ...)
}
\arguments{
\item{x}{SpatRaster. A presence-absence SpatRaster with the layers ordered according to the tree order.}

\item{branch.length}{numeric. A Named numeric vector containing the branch length of each specie.}

\item{filename}{character. Output filename.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculate phylogenetic diversity using rasters as input and output. This function follows Faith (1992).
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phylogrid"))
data <- phylo.pres(ras, tree)
range <- inv.range(data$x, data$branch.length)
rast.pd(data$x, data$branch.length)
}
}
\references{
Faith, D. P. (1992). Conservation evaluation and phylogenetic diversity. Biological conservation, 61(1), 1-10.
}
