% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pd.ses.R
\name{rast.pd}
\alias{rast.pd}
\title{Calculate phylogenetic diversity for raster data}
\usage{
rast.pd(x, branch.length, cores = 1, filename = "", ...)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species. The layers (species) must be sorted according to the tree order. See the phylo.pres function.}

\item{branch.length}{numeric. A Named numerical vector containing the branch length for a set of species.}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{filename}{character. Output filename.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculate the sum of the branch length for species present in each cell of the raster.
}
\examples{
library(terra)
library(phyloraster)
x <- rast(system.file("extdata", "rast.presab.tif", package="phyloraster"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phyloraster"))
data <- phylo.pres(x, tree)
pd <- rast.pd(data$x, data$branch.length)
plot(pd)
}
\references{
Faith, D. P. (1992). Conservation evaluation and phylogenetic diversity. Biological conservation, 61(1), 1-10.
}
\author{
Neander Marcel Heming and Gabriela Alves-Ferreira
}
