% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.null.mod.R
\name{rast.pd.ses}
\alias{rast.pd.ses}
\title{Phylogenetic diversity - Null model}
\usage{
rast.pd.ses(pres.reord, branch.length, aleats)
}
\arguments{
\item{pres.reord}{SpatRaster. A presence-absence SpatRaster with the layers ordered according to the tree order}

\item{branch.length}{numeric. A Named numeric vector containing the branch length of each specie}

\item{aleats}{numeric. A vector containing the number of aleatorizations}
}
\value{
SpatRaster
}
\description{
Phylogenetic diversity - Null model
}
\references{
Faith, D. P. (1992). Conservation evaluation and phylogenetic diversity. Biological conservation, 61(1), 1-10.
}
