% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pd.ses.R
\name{rast.pd.ses}
\alias{rast.pd.ses}
\title{Phylogenetic diversity (PD. Faith, 1992) standardized for specie richness}
\usage{
rast.pd.ses(
  x,
  branch.length,
  aleats,
  random = c("tip", "site", "specie", "fullspat"),
  filename = NULL
)
}
\arguments{
\item{x}{SpatRaster. A presence-absence SpatRaster.}

\item{branch.length}{numeric. A Named numeric vector containing the branch length of each specie. See phylo.pres function.}

\item{aleats}{numeric. A number indicating how many times the calculation should be repeated.}

\item{random}{character. A character indicating what type of randomization. Could be by tip, site, specie or full spat(site and specie).}

\item{filename}{character. Output filename.}
}
\value{
SpatRaster
}
\description{
The function calculates the PD corrected for species richness. In each null
model run, richness is kept constant and branch lengths of each species are
randomized. The function provides the mean, standard deviation of all null
models and also calculates the standardized effect size (SES).
}
\references{
Faith, D. P. (1992). Conservation evaluation and phylogenetic diversity. Biological conservation, 61(1), 1-10.
}
\author{
Gabriela Alves-Ferreira and Neander Heming
}
