% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pd.ses.R
\name{rast.pd.ses}
\alias{rast.pd.ses}
\title{Phylogenetic diversity standardized for species richness}
\usage{
rast.pd.ses(
  x,
  branch.length,
  aleats,
  random = c("tip", "site", "species", "both"),
  cores = 1,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species.}

\item{branch.length}{numeric. A named numerical vector containing the branch length of each species. See phylo.pres function.}

\item{aleats}{positive integer. A positive integer indicating how many times the calculation should be repeated.}

\item{random}{character. A character indicating the type of randomization. The currently available randomization methods are "tip", "site", "species" or "both" (site and species).}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{filename}{character. Output filename.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster. The function returns the observed phylogenetic diversity, the mean of the simulations calculated over n times, the standard deviation of the simulations, and the standardized effect size (SES) for the phylogenetic diversity.
}
\description{
Calculates the standardized effect size for phylogenetic diversity. The function has four different methods for spatial and phylogenetic randomization. See Details for more information.
}
\details{
The "tip" method shuffles the taxon names among all those in the phylogeny. The "site" method keeps species richness constant in each pixel, but randomizes the position of species in the stack. In the "species" method, the order of species in the stack is kept constant, but the pixels where each species is present are randomized in space. The third method, "full.spat", combines site and species randomization at the same time.
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phylogrid"))
data <- phylogrid::phylo.pres(ras, tree)
t <- phylogrid::rast.pd.ses(data$x, data$branch.length, aleats = 10, random = "species")
plot(t)
}
}
\references{
Faith, D. P. (1992). Conservation evaluation and phylogenetic diversity. Biological conservation, 61(1), 1-10.
}
\author{
Gabriela Alves-Ferreira and Neander Heming
}
