% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{rast.pe}
\alias{rast.pe}
\title{Calculate phylogenetic endemism (PE. Rosauer et al. 2009) for a raster}
\usage{
rast.pe(x, branch.length, filename = NULL, cores = 1, ...)
}
\arguments{
\item{x}{SpatRaster. A presence-absence SpatRaster with the layers ordered according to the tree order}

\item{branch.length}{numeric. A Named numeric vector containing the branch length of each specie}

\item{filename}{character. Output filename.}

\item{cores}{numeric. A positive integer indicating the number of clusters used for parallelization}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculate phylogenetic endemism following Rosauer et al. (2009) using rasters as input and output.
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phylogrid"))
data <- phylo.pres(ras, tree)
rast.pe(data$x, data$branch.length)
}
}
\references{
Rosauer, D. A. N., Laffan, S. W., Crisp, M. D., Donnellan, S. C., & Cook, L. G. (2009). Phylogenetic endemism: a new approach for identifying geographical concentrations of evolutionary history. Molecular ecology, 18(19), 4061-4072.
}
