% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pe.ses.R
\name{rast.pe.ses}
\alias{rast.pe.ses}
\title{Phylogenetic endemism standardized for specie richness}
\usage{
rast.pe.ses(
  x,
  branch.length,
  aleats,
  random = c("area.size", "site", "specie", "fullspat"),
  filename = NULL
)
}
\arguments{
\item{x}{SpatRaster. A presence-absence SpatRaster with the layers ordered according to the branch.length order. See phylo.pres function.}

\item{branch.length}{numeric. A Named numeric vector containing the branch length of each specie. See phylo.pres function.}

\item{aleats}{numeric. A number indicating how many times the calculation should be repeated.}

\item{random}{character. A character indicating what type of randomization. Could be by tip, site, specie or full spat(site and specie).}

\item{filename}{character. Output filename.}
}
\value{
SpatRaster
}
\description{
The function calculates the PE corrected for species richness. In each null model run, richness is kept constant and branch lengths of each species are randomized. The function provides the mean and standard deviation of all null models and also calculates the standardized effect size (SES).
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phylogrid"))
data <- phylo.pres(ras, tree)
t <- rast.pe.ses(data$x, data$branch.length, aleats = 10, random = "fullspat")
plot(t)
}

}
\references{
Rosauer, D. A. N., Laffan, S. W., Crisp, M. D., Donnellan, S. C., & Cook, L. G. (2009). Phylogenetic endemism: a new approach for identifying geographical concentrations of evolutionary history. Molecular ecology, 18(19), 4061-4072.
}
\author{
Gabriela Alves-Ferreira and Neander Heming
}
