% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.pe.ses.R
\name{rast.pe.ses}
\alias{rast.pe.ses}
\title{Phylogenetic endemism standardized for specie richness}
\usage{
rast.pe.ses(
  x,
  branch.length,
  aleats,
  random = c("tip", "spat"),
  cores = 1,
  filename = "",
  ...
)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species.}

\item{branch.length}{numeric. A named numerical vector containing the branch length of each species. See phylo.pres function.}

\item{aleats}{positive integer. A positive integer indicating how many times the calculation should be repeated.}

\item{random}{character. A character indicating the type of randomization. The currently available randomization methods are "tip", "site", "species" or "both" (site and species).}

\item{cores}{positive integer. If cores > 1, a 'parallel' package cluster with that many cores is created and used.}

\item{filename}{character. Output filename.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculates the standardized effect size for phylogenetic endemism. See Details for more information.
}
\details{
The spatial randomization (spat) keeps the richness exact and samples species presences proportionally to their observed frequency (i.e. number of occupied pixels). The randomization will not assign values to cells with nodata. The phylogenetic randomization shuffles taxa names across all taxa included in phylogeny.
}
\examples{
\dontrun{
x <- terra::rast(system.file("extdata", "rast.presab.tif", package="phyloraster"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phyloraster"))
data <- phylo.pres(x, tree)
t <- rast.pe.ses(data$x, data$branch.length, aleats = 10, random = "spat")
plot(t)
}
}
\references{
Rosauer, D. A. N., Laffan, S. W., Crisp, M. D., Donnellan, S. C., & Cook, L. G. (2009). Phylogenetic endemism: a new approach for identifying geographical concentrations of evolutionary history. Molecular ecology, 18(19), 4061-4072.
}
\author{
Gabriela Alves-Ferreira and Neander Heming
}
