% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.phylo.R
\name{rast.we}
\alias{rast.we}
\title{Calculate Weighted Endemism (WE) for a raster}
\usage{
rast.we(x, range.size, filename = NULL, ...)
}
\arguments{
\item{x}{SpatRaster. A presence-absence SpatRaster with 0 representing absence and 1 representing presence.}

\item{range.size}{Named numeric vector. A vector containing range size. See the function phylogrid::range.size.}

\item{filename}{character. Output filename.}

\item{...}{additional arguments to be passed passed down from a calling function.}
}
\value{
SpatRaster
}
\description{
Calculate weighted endemism using rasters as input and output.
}
\examples{
\dontrun{
ras <- terra::rast(system.file("extdata", "rast.presab.tif", package="phylogrid"))
rs <- phylogrid::range.size(ras)
rast.we(ras, rs)
}
}
\references{
Williams, P.H., Humphries, C.J., Forey, P.L., Humphries, C.J., VaneWright, R.I. (1994). Biodiversity, taxonomic relatedness, and endemism in conservation. In: Systematics and Conservation Evaluation (eds Forey PL, Humphries CJ, Vane-Wright RI), p. 438. Oxford University Press, Oxford.

Crisp, M., Laffan, S., Linder, H., Monro, A. (2001). Endemism in theAustralian flora. Journal of Biogeography, 28, 183–198.
}
