% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp2rast.R
\name{shp2rast}
\alias{shp2rast}
\title{Rasterize shapefile}
\usage{
shp2rast(shp, sps)
}
\arguments{
\item{shp}{SpatialPolygonsDataFrame or a SpatVector. A shapefile representing species distribution}

\item{sps}{vector. A vector of characters with unique (not duplicated) species names presents in the shapefile}
}
\value{
SpatRaster
}
\description{
The function transform a shapefile to a raster with the same extenction.
}
