% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.rand.R
\name{tip.rand}
\alias{tip.rand}
\title{Randomize branch lengths for n times}
\usage{
tip.rand(x, aleats, replace = FALSE)
}
\arguments{
\item{x}{numeric. A numeric vector containing the branch lengths.}

\item{aleats}{positive integer. A positive integer indicating how many times the calculation should be repeated.}

\item{replace}{should sampling be with replacement?}
}
\value{
list of numeric vectors
}
\description{
This function randomize branch lengths or another vector for n times.
}
\examples{
x <- terra::rast(system.file("extdata", "rast.presab.tif", package="phyloraster"))
tree <- ape::read.tree(system.file("extdata", "tree.nex", package="phyloraster"))
data <- phylo.pres(x = x, tree = tree)
sa <- tip.rand(data$branch.length, aleats = 10)


}
\author{
Gabriela Alves-Ferreira
}
